Validate utf codepoint at full width, not after 32-bit truncation
<<N/utf8>> (and utf16/utf32) must raise badarg when N is not a valid
Unicode code point. AtomVM rejected most invalid values but silently
encoded a family of large negative integers: the opcode handlers and JIT
primitives computed a 64-bit avm_int_t src_value, then passed it to
encoders whose codepoint parameter was uint32_t (and to the JIT
primitives, whose parameter was int). The value was implicitly truncated
to its low 32 bits before is_invalid_codepoint(int32_t) ever saw it. A
value such as -2^59 (0xF800_0000_0000_0000) truncates to 0 -- a valid
codepoint -- so <<(-576460752303423488)/utf8>> produced <<0>> where BEAM
raises badarg; -2^59+65 produced <<"A">>. This affected both the emulator
and JIT builds (erlfuzz bug #16, utf_negative_bignum_bypass).

Widen the codepoint all the way through: is_invalid_codepoint now takes
int64_t, the bitstring_utf{8,16,32}_encode / _size / insert_utf{8,16,32}
functions take int64_t c, and the jit.c primitives + jit.h interface
pointers match. The full-width value now reaches the range check, so
out-of-range negatives (and any value outside [0, 0x10FFFF]) are rejected
on both builds. The encoders only dereference c after the range gate, so
the narrower arithmetic inside them is unchanged. No Erlang codegen
change is needed -- the JIT already passes a full register-width value;
only the C declared parameter types were narrowing it.

Regression test in test_bs_utf: put_utf8/16/32 now assert badarg for
-2^59 and -2^59+65 (low 32 bits 0 and 65). Verified on JIT and emulator.

(cherry picked from commit 9bc3b54a520c83a72c964be99db1c1e39a9cb53e)

+++ src/libAtomVM/bitstring.c
@@ -144,7 +144,7 @@ bool bitstring_insert_any_integer(uint8_t *dst, avm_int_t offset, avm_int64_t va
     return true;
 }
 
-static bool is_invalid_codepoint(int32_t v)
+static bool is_invalid_codepoint(int64_t v)
 {
     return (v < 0) || (v > 0x10FFFF) || (v >= 0xD800 && v <= 0xDFFF);
 }
@@ -165,7 +165,7 @@ static bool is_invalid_codepoint(int32_t v)
 // +----------+----------+----------+----------+----------+
 //
 
-bool bitstring_utf8_encode(uint32_t c, uint8_t *buf, size_t *out_size)
+bool bitstring_utf8_encode(int64_t c, uint8_t *buf, size_t *out_size)
 {
     size_t sz = 0;
     if (is_invalid_codepoint(c)) {
@@ -208,7 +208,7 @@ bool bitstring_utf8_encode(uint32_t c, uint8_t *buf, size_t *out_size)
 //  W1 = 110110yyyyyyyyyy      // 0xD800 + yyyyyyyyyy
 //  W2 = 110111xxxxxxxxxx      // 0xDC00 + xxxxxxxxxx
 
-bool bitstring_utf16_encode(uint32_t c, uint8_t *buf, enum BitstringFlags bs_flags, size_t *out_size)
+bool bitstring_utf16_encode(int64_t c, uint8_t *buf, enum BitstringFlags bs_flags, size_t *out_size)
 {
     size_t sz = 0;
     if (is_invalid_codepoint(c)) {
@@ -303,7 +303,7 @@ bool bitstring_utf16_decode(const uint8_t *buf, size_t len, int32_t *c, size_t *
     return false;
 }
 
-bool bitstring_utf32_encode(uint32_t c, uint8_t *buf, enum BitstringFlags bs_flags)
+bool bitstring_utf32_encode(int64_t c, uint8_t *buf, enum BitstringFlags bs_flags)
 {
     UNUSED(bs_flags);
     if (is_invalid_codepoint(c)) {
